requirements.txt
blinker==1.9.0
click==8.3.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.1
gunicorn==26.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.9
mysql-connector-python==9.7.0
numpy==2.4.4
packaging==26.2
pandas==3.0.2
pillow==12.2.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
threadpoolctl==3.6.0
tzdata==2026.2
Werkzeug==3.1.8
